Add a-a test
